requirements.txt
feedparser
pyyaml
kafka-python
